Exclude fuzz testing from pre-1.18 toolchains
